Complete sorting re-write
Data is no longer sorted.  It is stored in a ring buffer and maps are used to determine the location of the median sample.  With this change, filtering is faster.  In the future, adapting the filter to use other data types (longs or floats) should have minimal effect on performance.